Fix ga frontend package updater (#64)
This pull request makes a minor update to the workflow permissions for
the dependency update automation. The change adds read permissions for
checks and actions to ensure the workflow has access to necessary
resources.

* Workflow permissions: Added `checks: read` and `actions: read` to the
`.github/workflows/update-deps.yml` file to allow the workflow to read
check results and actions data.